split ffi files
