+++ ddtelemetry-ffi/src/lib.rs
@@ -1,14 +1,13 @@
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present Datadog, Inc.
 
-use std::path::{Path, PathBuf};
-
 use ddcommon_ffi as ffi;
-use ddtelemetry::worker::{TelemetryWorkerBuilder, TelemetryWorkerHandle};
+use ddtelemetry::worker::TelemetryWorkerHandle;
 use ffi::slice::AsBytes;
 
-#[cfg(unix)]
-pub mod unix;
+pub mod builder;
+pub mod sidecar;
+pub mod worker_handle;
 
 macro_rules! c_setters {
     (
@@ -29,7 +28,7 @@ macro_rules! c_setters {
                     $object_name: &mut $object_ty,
                     param: $property_type,
                 ) -> MaybeError {
-                    $object_name . $path $(.  $path_rest)* = Some(try_c!($convert_fn (param)));
+                    $object_name . $path $(.  $path_rest)* = Some(crate::try_c!($convert_fn (param)));
                     MaybeError::None
                 }
             )+
@@ -57,7 +56,7 @@ macro_rules! c_setters {
                 match property {
                     $(
                         [< $path:camel $($path_rest:camel)* >] => {
-                            $object_name . $path $(.  $path_rest)* = Some(try_c!($convert_fn (param)));
+                            $object_name . $path $(.  $path_rest)* = Some(crate::try_c!($convert_fn (param)));
                         }
                     )+
                 }
@@ -79,11 +78,11 @@ macro_rules! c_setters {
                 property: ffi::CharSlice,
                 param: $property_type,
             ) -> MaybeError {
-                let property = try_c!(property.try_to_utf8());
+                let property = crate::try_c!(property.try_to_utf8());
                 match property {
                     $(
                         stringify!($path $(. $path_rest)*) => {
-                            $object_name . $path $(.  $path_rest)* = Some(try_c!($convert_fn (param)));
+                            $object_name . $path $(.  $path_rest)* = Some(crate::try_c!($convert_fn (param)));
                         }
                     )+
                     // TODO this is an error
@@ -96,12 +95,6 @@ macro_rules! c_setters {
     };
 }
 
-type MaybeError = ffi::Option<ffi::Vec<u8>>;
-
-#[no_mangle]
-pub extern "C" fn ddog_MaybeError_drop(_: MaybeError) {}
-
-#[macro_export]
 macro_rules! try_c {
     ($failable:expr) => {
         match $failable {
@@ -111,234 +104,18 @@ macro_rules! try_c {
     };
 }
 
-/// # Safety
-/// * builder should be a non null pointer to a null pointer to a builder
-#[no_mangle]
-pub unsafe extern "C" fn ddog_builder_instantiate(
-    builder: &mut *mut TelemetryWorkerBuilder,
-    service_name: ffi::CharSlice,
-    language_name: ffi::CharSlice,
-    language_version: ffi::CharSlice,
-    tracer_version: ffi::CharSlice,
-) -> MaybeError {
-    let new = Box::new(TelemetryWorkerBuilder::new_fetch_host(
-        service_name.to_utf8_lossy().into_owned(),
-        language_name.to_utf8_lossy().into_owned(),
-        language_version.to_utf8_lossy().into_owned(),
-        tracer_version.to_utf8_lossy().into_owned(),
-    ));
-    // Leaking is the last thing we do before returning
-    // Otherwise we would need to manually drop it in case of error
-    *builder = Box::into_raw(new);
-    MaybeError::None
-}
+pub(crate) use c_setters;
+pub(crate) use try_c;
 
-/// # Safety
-/// * builder should be a non null pointer to a null pointer to a builder
-#[no_mangle]
-pub unsafe extern "C" fn ddog_builder_instantiate_with_hostname(
-    builder: &mut *mut TelemetryWorkerBuilder,
-    hostname: ffi::CharSlice,
-    service_name: ffi::CharSlice,
-    language_name: ffi::CharSlice,
-    language_version: ffi::CharSlice,
-    tracer_version: ffi::CharSlice,
-) -> MaybeError {
-    let new = Box::new(TelemetryWorkerBuilder::new(
-        hostname.to_utf8_lossy().into_owned(),
-        service_name.to_utf8_lossy().into_owned(),
-        language_name.to_utf8_lossy().into_owned(),
-        language_version.to_utf8_lossy().into_owned(),
-        tracer_version.to_utf8_lossy().into_owned(),
-    ));
-
-    // Leaking is the last thing we do before returning
-    // Otherwise we would need to manually drop it in case of error
-    *builder = Box::into_raw(new);
-    MaybeError::None
-}
-
-c_setters!(
-    object_name => builder,
-    object_type => TelemetryWorkerBuilder,
-    property_type => ffi::CharSlice,
-    property_type_name_snakecase => str,
-    property_type_name_camel_case => Str,
-    convert_fn => (|s: ffi::CharSlice| -> Result<_, String> { Ok(s.to_utf8_lossy().into_owned()) }),
-    SETTERS {
-        application.service_version,
-        application.env,
-        application.runtime_name,
-        application.runtime_version,
-        application.runtime_patches,
-
-        host.container_id,
-        host.os,
-        host.kernel_name,
-        host.kernel_release,
-        host.kernel_version,
-
-        runtime_id
-    }
-);
-
-c_setters!(
-    object_name => builder,
-    object_type => TelemetryWorkerBuilder,
-    property_type => bool,
-    property_type_name_snakecase => bool,
-    property_type_name_camel_case => Bool,
-    convert_fn => (|b: bool| -> Result<_, String> { Ok(b) }),
-    SETTERS {
-        config.telemetry_debug_logging_enabled,
-    }
-);
-
-c_setters!(
-    object_name => builder,
-    object_type => TelemetryWorkerBuilder,
-    property_type => ffi::CharSlice,
-    property_type_name_snakecase => path,
-    property_type_name_camel_case => Path,
-    convert_fn => (|p: ffi::CharSlice| -> Result<PathBuf, String> { Ok(Path::new(p.to_utf8_lossy().as_ref()).to_path_buf()) }),
-    SETTERS {
-        config.mock_client_file,
-    }
-);
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_builder_with_native_deps(
-    builder: &mut TelemetryWorkerBuilder,
-    include_native_deps: bool,
-) -> MaybeError {
-    builder.native_deps = include_native_deps;
-    MaybeError::None
-}
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_builder_with_rust_shared_lib_deps(
-    builder: &mut TelemetryWorkerBuilder,
-    include_rust_shared_lib_deps: bool,
-) -> MaybeError {
-    builder.rust_shared_lib_deps = include_rust_shared_lib_deps;
-    MaybeError::None
-}
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_builder_with_config(
-    builder: &mut TelemetryWorkerBuilder,
-    name: ffi::CharSlice,
-    value: ffi::CharSlice,
-) -> MaybeError {
-    let name = name.to_utf8_lossy().into_owned();
-    let value = value.to_utf8_lossy().into_owned();
-    builder.library_config.push((name, value));
-    MaybeError::None
-}
+type MaybeError = ffi::Option<ffi::Vec<u8>>;
 
 #[no_mangle]
-/// # Safety
-/// * handle should be a non null pointer to a null pointer
-pub unsafe extern "C" fn ddog_builder_run(
-    builder: Box<TelemetryWorkerBuilder>,
-    handle: &mut *mut TelemetryWorkerHandle,
-) -> MaybeError {
-    *handle = Box::into_raw(Box::new(try_c!(builder.run())));
-    MaybeError::None
-}
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_handle_add_dependency(
-    handle: &TelemetryWorkerHandle,
-    dependency_name: ffi::CharSlice,
-    dependency_version: ffi::CharSlice,
-) -> MaybeError {
-    let name = dependency_name.to_utf8_lossy().into_owned();
-    let version = dependency_version
-        .is_empty()
-        .then(|| dependency_version.to_utf8_lossy().into_owned());
-    try_c!(handle.add_dependency(name, version));
-    MaybeError::None
-}
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_handle_add_integration(
-    handle: &TelemetryWorkerHandle,
-    dependency_name: ffi::CharSlice,
-    dependency_version: ffi::CharSlice,
-    compatible: ffi::Option<bool>,
-    enabled: ffi::Option<bool>,
-    auto_enabled: ffi::Option<bool>,
-) -> MaybeError {
-    let name = dependency_name.to_utf8_lossy().into_owned();
-    let version = dependency_version
-        .is_empty()
-        .then(|| dependency_version.to_utf8_lossy().into_owned());
-    try_c!(handle.add_integration(
-        name,
-        version,
-        compatible.into(),
-        enabled.into(),
-        auto_enabled.into(),
-    ));
-    MaybeError::None
-}
-
-#[allow(clippy::missing_safety_doc)]
-#[no_mangle]
-pub unsafe extern "C" fn ddog_handle_add_log(
-    handle: &TelemetryWorkerHandle,
-    indentifier: ffi::CharSlice,
-    message: ffi::CharSlice,
-    level: ddtelemetry::data::LogLevel,
-    stack_trace: ffi::CharSlice,
-) -> MaybeError {
-    try_c!(handle.add_log(
-        indentifier.as_bytes(),
-        message.to_utf8_lossy().into_owned(),
-        level,
-        stack_trace
-            .is_empty()
-            .then(|| stack_trace.to_utf8_lossy().into_owned()),
-    ));
-    MaybeError::None
-}
-
-#[no_mangle]
-pub extern "C" fn ddog_handle_start(handle: &TelemetryWorkerHandle) -> MaybeError {
-    try_c!(handle.send_start());
-    MaybeError::None
-}
-
-#[no_mangle]
-pub extern "C" fn ddog_handle_clone(handle: &TelemetryWorkerHandle) -> Box<TelemetryWorkerHandle> {
-    Box::new(handle.clone())
-}
-
-#[no_mangle]
-pub extern "C" fn ddog_handle_stop(handle: &TelemetryWorkerHandle) -> MaybeError {
-    try_c!(handle.send_stop());
-    MaybeError::None
-}
-
-#[no_mangle]
-pub extern "C" fn ddog_handle_wait_for_shutdown(handle: Box<TelemetryWorkerHandle>) {
-    handle.wait_for_shutdown()
-}
-
-#[no_mangle]
-pub extern "C" fn ddog_handle_drop(handle: Box<TelemetryWorkerHandle>) {
-    drop(handle);
-}
+pub extern "C" fn ddog_MaybeError_drop(_: MaybeError) {}
 
 #[cfg(test)]
 mod test_c_ffi {
     use super::*;
+    use crate::{builder::*, worker_handle::*};
 
     #[test]
     fn test_set_builder_mock_client_config() {
