+++ ddtelemetry-ffi/src/sidecar/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(unix)]
+pub mod unix;
