+++ ddtelemetry-ffi/src/sidecar/unix.rs
